refactor: split out mod removal to its own workflow
