+++ workflows/types.go
@@ -1,6 +1,7 @@
 package workflows
 
 import (
+	"github.com/satisfactorymodding/smr-api/workflows/removemod"
 	"github.com/satisfactorymodding/smr-api/workflows/statistics"
 	"github.com/satisfactorymodding/smr-api/workflows/updatemodfromstorage"
 	"github.com/satisfactorymodding/smr-api/workflows/versionupload"
@@ -10,8 +11,10 @@ var Workflows = struct {
 	Statistics           *statistics.A
 	UpdateModFromStorage *updatemodfromstorage.A
 	VersionUpload        *versionupload.A
+	RemoveMod            *removemod.A
 }{
 	Statistics:           statistics.Statistics,
 	UpdateModFromStorage: updatemodfromstorage.UpdateModFromStorage,
 	VersionUpload:        versionupload.VersionUpload,
+	RemoveMod:            removemod.RemoveMod,
 }
